refactor(apex-debugger): move sole-consumer commandlet framework local
Move PostconditionChecker, PreconditionChecker, CompositeParametersGatherer,
and EmptyParametersGatherer from salesforcedx-utils-vscode into apex-debugger's
sfCommandlet.ts, their only consumer. Delete dead SfCommandletExecutor and
LibraryCommandletExecutor (no consumers; apex-debugger uses core's runtime
export). Keeps the whole commandlet framework in one place instead of
cross-package imports for single-consumer symbols.
