+++ packages/salesforcedx-utils-vscode/src/commands/parameterGatherers.ts
@@ -18,33 +18,3 @@ export type CancelResponse = {
 export type ParametersGatherer<T> = {
   gather(): Promise<CancelResponse | ContinueResponse<T>>;
 };
-
-export class CompositeParametersGatherer<T> implements ParametersGatherer<T> {
-  private readonly gatherers: ParametersGatherer<any>[];
-  constructor(...gatherers: ParametersGatherer<any>[]) {
-    this.gatherers = gatherers;
-  }
-  public async gather(): Promise<CancelResponse | ContinueResponse<T>> {
-    const aggregatedData: any = {};
-    for (const gatherer of this.gatherers) {
-      const input = await gatherer.gather();
-      if (input.type === 'CONTINUE') {
-        Object.keys(input.data).map(key => (aggregatedData[key] = input.data[key]));
-      } else {
-        return {
-          type: 'CANCEL'
-        };
-      }
-    }
-    return {
-      type: 'CONTINUE',
-      data: aggregatedData
-    };
-  }
-}
-
-export class EmptyParametersGatherer implements ParametersGatherer<{}> {
-  public gather(): Promise<CancelResponse | ContinueResponse<{}>> {
-    return Promise.resolve({ type: 'CONTINUE', data: {} });
-  }
-}
