--- packages/salesforcedx-utils-vscode/src/commands/postconditionCheckers.ts
@@ -1,11 +1,0 @@
-/*
- * Copyright (c) 2026, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-import type { CancelResponse, ContinueResponse } from './parameterGatherers';
-
-export type PostconditionChecker<T> = {
-  check(inputs: ContinueResponse<T> | CancelResponse): Promise<ContinueResponse<T> | CancelResponse>;
-};
