--- packages/salesforcedx-utils-vscode/src/commands/preconditionCheckers.ts
@@ -1,10 +1,0 @@
-/*
- * Copyright (c) 2026, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-
-export type PreconditionChecker = {
-  check(): Promise<boolean> | boolean;
-};
